requirements.txt
acachecontrol==0.3.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
arcadepy==1.4.0
asgiref==3.8.1
attrs==25.3.0
Authlib==1.5.2
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.1
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.3
decorator==5.2.1
distro==1.9.0
Django==5.1.7
ecdsa==0.19.1
feedparser==6.0.11
firecrawl-py==2.5.4
Flask==3.0.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flower==2.0.1
frozenlist==1.6.0
google-ai-generativelanguage==0.6.6
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
judoscale==1.9.0
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
okta-jwt-verifier==0.2.8
openai==1.78.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
retry2==0.9.5
rsa==4.9.1
setuptools==80.4.0
sgmllib3k==1.0.0
six==1.17.0
slack_sdk==3.27.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.28
SQLAlchemy-serializer==1.4.1
sqlparse==0.5.3
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
